[FIX] 목표 설정 session 오류 해결 #62
